services/agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agents-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.2",
    "google-adk>=0.0.6",
    "google-genai>=1.46.0",
    "pydantic>=2.8.2",
    "pydantic-settings>=2.5.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "fastembed>=0.4.0",
    "qdrant-client>=1.11.1",
    "requests>=2.32.3",
    "ruff>=0.14.2",
    "uvicorn[standard]>=0.30.6",
    "pytest>=7.4.4",
    "respx>=0.20.2",
    "httpx>=0.27.0",
]
